pyproject.toml
[project]
dependencies = ["clicK", "pyyaml", "pony", "rich"]                              
name = "flug"
requires-python = ">= 3.11"
scripts = { flug = "flug:cli" } 
version = "0.1.0"

[build-system] 
build-backend = "hatchling.build"
requires = ["hatchling"]